Book-physical-ai/rag-chatbot/backend/requirements.txt
# AI Assistant Backend Dependencies
# Install with: pip install -r requirements.txt

# FastAPI and ASGI server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# CORS support
python-multipart>=0.0.6

# Pydantic for data validation
pydantic>=2.5.0

# LangChain for conversational AI and RAG orchestration
langchain>=0.2.0
langchain-google-genai>=1.0.0
langchain-community>=0.2.0
langchain-core>=0.2.0

# ChromaDB for vector storage
chromadb>=0.5.0

# Google Generative AI for embeddings and LLM
google-generativeai>=0.8.0

# Python environment management
python-dotenv>=1.0.0

# Optional: For markdown processing
markdown>=3.5.0
